fix: Improved controller handshaking with the mod
Due to the way the handshaking used to work, the Arduino would get stuck
waiting for serial data that would never arrive, as the mod had not seen
any handshake attempts to respond to. This fix simple adds in a retry
system that will check for available serial data a few times before
returning that the init did not work - which should now mean that the
way the simpit init is demonstrated in the demo code will actually
work as intended.