Switched readme from rst to md, and minor rewrite
I'm of the personal opinion that markdown is the better choice, but I fully understand if a rst version would be preferred. The main purpose of this commit is to rewrite the readme to reflect the status it holds as a rewrite and link to the correct accompanying library, etc.